requirements.txt
# --- 原本的專案依賴 ---
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.2
pydantic_core==2.41.4
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0

# --- 新增：Firebase Admin (Auth 功能) ---
firebase-admin>=6.0.0

# --- 新增：RAG AI 小幫手功能 ---
langchain-openai>=0.3.35
langchain-core>=0.3.78
langchain-pinecone>=0.2.0
pinecone-client>=5.0.0